Making a number of changes to both the basis_sqrt and SR Gamma methods to bring them more in line with how similar operations are performed in the PKPM framework.
Now we more aggressively check the positivity at control points of (1 - V^2/c^2) and further inside the sqrt operation, we set the quadrature point to be 0 if it detects the evaluation of the function at that quadrature point is negative (to avoid the sqrt returning a NaN).

This set of robustness checking reveals there's still an issue with correcting all the moments in the SR BGK test, as the correction after the first time step seems to drive the distribution function's moments to NaN, even the drift velocity (which is why the Gamma_inv division fails with correct_all_moms turned on).