Further robustification of the relativistic Gamma factor from the bulk velocity. The problem was (that the correct routine nicely picked out) if a cell is undergoing rapid variation, then even if V_drift < c, V_drift^2 > c^2 (because in our DG expansion, the square of an expansion has, for example, for the cell average a sum of the squares of the coefficients, so large slopes can make for very large V_drift^2).
So we check inside the kernels if the cell average of V_drift^2 > c^2 (where c = 1 in all relativistic calculations). If it is, we know that the cell has large slopes and higher order moments, and we zero out all those higher order coefficients for the computation of the relativistic Gamma factor. We can now run the relativistic BGK Sod shock test correcting all the moments for all time.