Adding a new method to basis called "quad_nodal_to_modal" which stores kernels that compute the one-to-one transformation of the Gauss-Legendre quadrature nodal basis (the nodes are Gauss-Legendre abscissas) and our modal basis.
In some sense, for many cases this supersedes the upwind_quad_to_modal functions, but I'm not currently trying to rewrite any equation kernels to use this. Instead I am using this as a test for improving the efficience of the projection routine for the Vlasov LTE distribution function since these kernels unroll all the loops for this conversion (where the loop was previously just a naive matrix-vector multiplication).

The efficiency gain is not large (10 percent) on CPUs. I will test on GPUs to see if the GPU efficiency gain is larger.