First pass at the implicit source solve in DG Fluids. Somehow this is "close" but in a way that doesn't make a whole lot of sense.
The way I approached the implicit source solve is to use the same ideas that go into weak division/the LBO primitive variable solve to do everything in the modal orthonormal basis. I construct the matrix system for our implicit source solve, solving the matrix system

 * J_s^{n+1} - 0.5*dt*(q_s^2/m_s^2*rho_s^n*E^{n+1} + q_s/m_s*J_s^{n+1} x B^n) = J_s^n + ext_forces
 * epsilon0*E^{n+1} + 0.5*dt*sum_s J_s^{n+1} = epsilon0*E^{n} - ext_currents

where the unknowns are J_s^{n+1} and epsilon0*E^{n+1} (these are actually the time-centered quantities so that the method is second order, but the idea is the same).

So all of these quantities have DG expansions, and we just use the DG expansion of rho (the mass density) and B (the magnetic field) at the known time steps so that the system is linear (since rho and B have no source terms in this construction).

The implementation of this required a refactor of the Vlasov app, following how things are done in the moments app where we have function pointers for the "update_func" and I separated out the forward Euler and SSP RK3 so that the Strang Split option was more transparent.

What's interesting is that this kind of works but seems to have some very serious bug in the logic for the initial time step (or time steps). For example, the mom_beach p=1 test starts developing NaNs where the antenna is, but if you run with p=2, it takes a couple DT_MIN junk time steps, and then starts taking regular time steps, and the result is that the mom_beach with p=2 looks very similar to the mom_beach for 5 moment through moment app. For the life of me I cannot figure out why this would happen but hopefully it means that this is genuinely "close"

**IMPORTANT** This work caused me to discover a bug in moment_em_coupling.c where there was an extra factor of 1/epsilon0 in the applied currents. I removed this as part of the comparisons between FV and DG.