Deleting some tests that do not actually function as regression tests right now. While progress has been made on DG fluids, this addition to the Vlasov app is fundamentally for a test bed for improvements to DG fluids *for* PKPM, and unless I implement a positivity fix for density and pressure, the 5-moment fluid solver will fail for tests like the Riemann problem (neutral Sod shock is fine, but once you add the EM fields things get crazy).
I am also deleting the mom_beach_p1 test. As I said in an earlier commit, DG fluids seem to work best when you run with high polynomial order + the limiters. p=2 and p=3 on coarse meshes reproduce the plasma beach, but p=1 crashes from dispersive errors (the limiters do not do enough it appears). Again, this limitation is not a high priority to "solve" at the moment because DG fluids are a testbed functionality within Gkeyll.