Because of recent discussions about wanting to explore positivity fixes within pure p=1, adding some pure p=1 functionality to Vlasov.
The goal here is *not* to supersede or rewrite any functionality. The default is hybrid when p=1 and for Vlasov we absolutely want p=2 in velocity space for components of the solve like the LBO.

However, I realized I could do something that was very philosophically satisfying that gives a desired testbed with the BGK collision operator + moments corrections. In the past I have wanted to support tensor basis functions as an option because I've had Vlasov calculations which are more robust with p=2 tensor basis functions. Now I just have the p=1 tensor basis be a *pure* p=1. This works great from the program flow perspective because now the tensor basis is wholly distinct from the Serendipity basis and the Serendipity basis, as the default option, can do whatever we want in terms of making it indirectly hybrid for p=1 and other default options we prefer to support.

Also this makes the Maxima nice from my perspective because now we support the same options for tensor and Serendipity (p=2 for 1x1v, 1x2v, 1x3v, 2x2v, 2x3v, and only p=1 for 3x3v).