Unfortunately right now we don't have a good way of passing an array of gkyl_arrays onto GPUs because this array of gkyl_arrays is not accessible on device (only the individual gkyl_arrays are).
As such, because all the multi-fluid and PKPM simulations are going to either be 1 fluid or 2-fluid for right now (at least during this testing phase) we will have two set and copy methods for packing and unpacking the matrix to compute the linear solve for the implicit fluid-EM coupling. Hopefully we can come up with a more elegant solution in time.