Adding the option to the gyrokinetic maxwellian correction updater to utilize the result of the iteration *even if it is not converged*. While such an option carries a risk that the moments are diverging, preliminary tests with the BGK sod shock is that, with large enough mu extents, the errors are dominantly in the parallel flow velocity and are difficult to converge to machine precision (for reasons I do not yet understand) but we can get 1.0e-14 errors in T/m (compared to 1 percent errors without the iterative scheme) in only 5-10 iterations.
Some further iteration (pun intended) is likely possible. I want to experiment with whether we can utilize a "converged" density and T/m but unconverged u_parallel to maybe get the best of both worlds, because in the process of converging in T/m we diverge slightly in u_parallel (the errors with no iterative scheme in u_parallel are ~1.0e-8-1.0e-5, but with the iteration that improves T/m our accuracy of u_parallel degrades).