Success! We can now converge all the moments in a BGK sod shock with the GK system in as few as 8 iterations to 1.0e-14 error.
The issue was that I had to switch to relative error for the density and temperature convergence because I was having serious finite precision errors trying to converge to temperature when using SI units. However, I (in quite silly fashion) also started using relative error for u_par. But u_par is 0 in some cells! So that's a bummer. Now I am using absolute error for u_par and everything works beautifully.