For proper functioning of the Lua layers, each function object in any App (moments, vlasov and GK) need to have their own function context. At present, we are assuming we have a common function context, which is ok for the C driver.
However, the way Lua-C layer works is that the reference to the Lua
function passed to the C layer is stored in a hidden context. This
means that if we use the same context for all functions in field or
species, then we will not be able to use different Lua functions! This
is not what we want, obviously. I am going through and changing this
but we need to do this for *every single function pointer passed to
all Apps*.