At MagNetUS, a discussion came up about if we can simulate LAPD and we have neutral effects, can we simulate the experiment at Auburn: MDPX (magnetized dusty plasma experiment). While MDPX is often a dusty plasma experiment, there are a number of studies that have been done in their setup that are pure plasma, but neutral dominated (neutral density 1 million times larger). I am adding a reference input file for this setup, which is basically a combination of the STEP input file with static neutrals and a Cartesian LAPD setup with identity mapping and uniform magnetic field. Going to try running on a couple of GPUs.
For right now, this input file runs for a very short period of time. MDPX is so short in the axial direction the transient evolution of the sheath sets a very restrictive time step but it seems like as the simulation runs this relaxes. Not sure if eventually we can overcome this transient but I am just brute forcing it for now.