Changing the implicit source solve to now perform an implicit update on du_s/dt instead of dJ_s/dt (where J_s = q_s/m_s rho_s u_s).
This change allows for an easier generalization to new terms such as inter-species collisions  (because otherwise there are non-trivial non-pre-computable divisions by the density within the implicit source solve). This formulation is also the way that implicit source solve will need to be constructed for, e.g., relativistic systems in which the only sensible implicit source solve is on du_s/dt (where now u_s is the four-velocity).

Everything I have tested still works. I have also tracked down that the differences in energy conservation I was seeing are because the implicit source solve is a lower order time stepping scheme than a fully explicit SSP RK3 time stepping scheme. This discovery motivates retaining the explicit source solve kernels in cases when the user is resolving all the scales. Need to do a little infrastructure work though to make that sensible.