I am officially making another attempt at formulating a part of the surface terms in PKPM utilizing weak operations *at the necessary surfaces*.
The essential idea is that the flow velocity u and magnetic field vector b are needed at surfaces. One can either compute a volume expansion of u and b and then evaluate that volume expansion at the surface, or instead evaluate the base quantities (rhou and rho in the case of u) at the surface *and then* perform the weak operations (so in 2D one would do 1D weak division at the needed surfaces).

Because all of our operations are not point-wise and only defined weakly, these two means of getting u (or b) at surfaces do not give equivalent values. Originally I abandoned this line of inquiry because I thought it was less robust. Now I am convinced that I was just not thinking about the operation correctly. The reality is that the way the first integration by parts is performed has some inherent freedom and we are exploting this freedom in doing weak operations at the surface. For example, when computing b, we can leverage the fact that we can retain *at least* the linear variation of b by computing things at needed surfaces in the case when B (the magnetic field) is rapidly varying and we have to revert to lower order expressions to avoid taking square roots of negative numbers.

In general this might be slightly less robust, but I do think it is more accurate and I can prove the desired properties I want (such as energy conservation and L2 stability) because I understand how the weak and strong forms manifest in this form. I think the recent change to the Riemann solve with Roe fluxes gets me 90+ percent of the way there for stable PKPM simulations.