Revert "I am officially making another attempt at formulating a part of the surface terms in PKPM utilizing weak operations *at the necessary surfaces*."
This reverts commit 9451511ab64036098b0866c9a6adf158e86c770b.

I am reverting this commit because this is not quite right. It turns out we *cannot* use u computed at surfaces and construct an energy conserving scheme because when you substitute u for the test function, it must be a valid test function *in the volume* that then is evaluated at surfaces when doing the proofs. I believe we can retain computing b at surfaces because b just exchanges energy between different degrees of freedom and thus there is more flexibility, but I want to revert this whole commit and re-organize so everything is cleaner.