Adding the ability for the correction routine to correct a BiMaxwellian, so that initial conditions for both Maxwellians and BiMaxwellians can be corrected to give the desired initial moments. Some additional cleanup of the logic in the projection object now that Maxwellians and BiMaxwellians share this functionality.
I have also added comments on the GPU specific function that is needed when applying the correction routine. Further I have tried to make the correction routine initialization reasonably agnostic to Maxwellian vs. BiMaxwellian, though for right now there is a maxwellian_correct and a bimaxwellian_correct. There may be a nice way to handle this generically so one doesn't need two functions, but this seems good enough for now.