Experimenting with a new formulation of PKPM in an attempt to derive and implement a truly energy-conserving scheme.
The difficulty lies in the evolution of the kinetic energy. In classical KEP schemes, the formulation of kinetic energy preservation lies in the test function for the mass density evolution being u^2/2 (the flow velocity squared). In a standard formulation of DG, this function is not in the basis because only u is in the basis (not u^2).

So we re-write the PKPM system from the ground up in a mixed basis formalism, where the kinetic equation is solved at order 2*p, while the momentum equation is solved at order p. Thus, in this formalism, rhou is order p, u is order p, and rho is order 2*p.

As far as I can tell from having Maxima do the algebra, this does genuinely create an energy conserving scheme. There is a unique operation for weak division to obtain u (because rho is just the weight of the inner product and thus can be of arbitrary order) since both u and rhou are the same order. Further the operation is invertible.

However, at a first pass implementing this scheme, the formulation is still not conservative. Preliminary regression testing suggests what I have done is "close" as certain tests are indistinguishable by eye between the old and new formulations. I would guess the likely culprit here is that there is some inconsistency in the construction of all the coupling terms, where something should be order 2*p but is only order p (or vice versa). The result also hinges on using tensor product bases at first glance because you need separability between the velocity space and configuration space bases to create the formalism (so that one can clearly extract the mass density and pressure evolution from the kinetic equation by substituting in the appropriate velocity test function).

The most difficult part of the proof hinges on whether the volume and surface contributions cancel for the u^2 weighting in the density evolution and u weighting of the momentum evolution. Again, with tensor bases we should be able to just plug in an equivalent mass density evolution to obtain this cancellation. But it's probably worth directly taking the evolution from the kinetic equation to see if there's some subtlety in the construction of the needed momentum fluxes from the mass flux. I also want to go back to directly computing div(P) now that the pressure tensor is order 2*p since it's obtained from the kinetic equation.

All the tests done thus far have static b, but an ostensible benefit of this formulation is that one can safely utilize a 2*p order basis for the construction of |B|^2 and thus guarantee positivity of |B|^2. But more work is needed to determine what the source of conservation errors is in this formalism before doing all the general testing previously done.