Discovery: the magnetic field unit vector in this mixed basis formalism *must* be p=1 (or the lower order basis in general). While bb is order 2*p now (which confers the added benefit that |B|^2 now) and originally I thought because the sqrt operation is nonlinear I would want b to also order 2*p, the only way to guarantee b . b = 1 is to represent b in the lower order basis.
There's still a subtlety that div(b) contains large variation if the positivity check activates (because bb < 0 at a quadrature point used to evaluate sqrt(bb) in the nodal basis). It's possible this is where the revival of the surface solve is needed (or that this approach just flat out does not work because it's phenomenally difficult to get a representation of b within these constraints).