Placeholder commit to update this branch to where my thinking ended up. I intend to use this branch as a prototyping branch as part of the new design review but my thinking has evolved on this front for the best way to handle this going forward.
Essentially I learned the lesson from re-checking my proofs that the hybrid mixed-basis formalism only applied to a single term in the update: the flow velocity u. Otherwise there is a down-projection of volume and surface updates from the mass flux that breaks the proof of kinetic energy conservation.

So the original idea that Maxwell's + the momentum equation would be p=1, while the kinetic equation would be p=2, does not have the desired properties. In fact, the *only* term which should be lower order is the flow velocity u, and then then quadratic rho*u is the quadratic rho weak multiplied the linear u.

Thus, the right design is to keep the PKPM system mostly as is, with an optional input boolean input parameter "kep" or something like that, which if the user specifies, then solves for the flow velocity at *lower order* leaving everything aspect of the update unchanged and throwing an assertion at a user that this input parameter does not make sense for p=1.