As part of the prototyping of the mixed basis formalism, I found increased robustness when utilizing the evaluation of rho*u at the interface for the Roe solve (so that when the 10 moment Roe solve computes the primitive variables, the Roe solve is genuinely utilizing the primitive variables at the interface; I believe this is unique to 10 moment because of how 10 moment utilizes the primitive variables to perform the Roe solve with the left and right eigenvectors multiplied by the phi' and phi'^{-1} matrices to convert between the two forms of the 10 moment system).
So I have switched the solve to utilize the primitive variables in this way (such that when the call to wv_ten_moment does the primitive variable reconstruction, it will get u and P at that quadrature point). More testing is needed but there do seem to be fewer oscillations in some of the initial tests performed.