Some clean up in preparation for the design review and hopeful merge.
Some careful testing suggested that the Strang split was not actually second order in time. My current understanding is the combination of the RK3 method, as a multi-stage method, with the time-centered method, somehow is different than the one-step method + Strang split moments uses. I do not know that this makes sense, but I know that the time-centered method on its own is second order, and the RK3 method on its own is third order, and when put together in the Strang split I was not getting second order accuracy.

So for now, deprecate the "strang split" update method and just put everything together for Vlasov and PKPM as an update_op_split method.