Some clean up of the regression tests which have been added within this branch. I discovered that part of the issue with the moment beach problem is that in DG, because of the higher order coefficients, the projection creates a negative density at x = Lx from the initial conditions (which are a (1 - x)^5 profile to simulate a very sharp density gradient that varies over orders of magnitude in the simulation).
So I add a small floor which amounts to a 10^-30 correction to the density but nevertheless makes the density flat near the edge of the domain. So the density still varies by ~30 orders of magnitude in the simulation, it just doesn't go negative at x = Lx.

Added a PKPM moment beach problem which seems to behave slightly different from the fluids case. The cutoff is in a slightly different location and you can actually see the waves damp when they go evanescent. Not sure if this is a bug or just extra physics in PKPM (I don't recall full Vlasov doing this, but I also do not closely recall how full Vlasov behaved).