Refactoring the SR Vlasov system of equations to construct a conservative DG formalism. Essential idea is that instead of directly projecting p/gamma (the relativistic velocity), we project gamma onto a continuous basis and then when p/gamma is needed, we do d/dp (gamma) (which in the continuous limit is p/gamma). Everywhere we need p/gamma (such as in streaming and in computing M1i) we take gamma as the input and take its derivatives. I have proved that this scheme is now energy conservative, with energy conservation (the gamma moment of the distribution) only depending on the size of the time step and converging at 3rd order with our RK3 scheme.
I have also added a radiation operator which just calls the LBO drag kernels with input collision frequency and drag coefficient (not a velocity in the relativistic case, but a four-velocity that the system relaxes too in momentum space).

I have also extended sources and initial conditions to take a number of input sources and projection objects, which allows us to utilize the Vlasov LTE object in relativity for all current regression tests. This refactor eliminates the normalization difficulties in relativity, even when doing Weibel or two-stream tests and makes the injected source in the escreen test more accurate.