Reorganizing Vlasov SR moment calculation and correction for robustness and generalizability.
The essential problem was that before, we were computing the drift velocity via weak division of M1i/M0 (like in the non-relativistic limit), but in regions of strong gradients in the velocity, the DG representation was often producing superluminal drift velocities, or even in the case where V_drift < c, V_drift^2 could be quite large.

So my proposed refactor is to take a two-step approach that first attempts to get a robust representation of the rest-frame density n, *and then* obtains the bulk four-velocity from weak division of n with M0 (to get GammaV) and M1i (to get GammaV*V_drift). My logic is that the initial distribution is projected with a re-scaling to have rest-frame density n, meaning that with respect to the rest-frame density the distribution function is fixed, so we just need to get some reasonably accurate representation of n and then everything else follows.

To get a reasonably accurate representation of n, we have the same algorithm as before that we compute V_drift as M1i/M0 with weak division, and we check if anywhere V_drift^2 > c^2. If it is, we switch to a p=1 Gauss-Lobatto representation of 1 - V_drift^2/c^2 at corners with a final robustness check of flooring this value to 1.0e-16 (so GammaV values of 1e8, which seem like a reasonable limit). Then we can guarantee n = GammaV_inv*M0, where GammaV_inv = sqrt(1 - V_drift^2/c^2) is well-defined everywhere in the cell.

The impact of this change is that a significant chunk of how we do moments can be reformulated using this more robust formulation. We now use this bulk four-velocity in the pressure computation and with relativity we correct (n, GammaV, GammaV*V_drift, T/m) (so vdim+3 components). I have also moved some things around for clarity of the API, for example, the "sr_vars" struct now handled the pressure computation so that all the mom_vlasov_sr moment type needs is gamma (the particle Lorentz boost factor), which greatly simplifies moment computation.

This refactor compiles but does not run yet because I am still refactoring the projection routine to use these more robust formulations of the bulk four-velocity, but I want to pull in the latest version of dg_10m that has the improvements to how we project the LTE distribution. I will also go through and try to clean up some of the naming conventions. If people have suggests for how to handle the difference between the spatial component of the particle four-velocity (which we consistently call p, even though it has no mass factor), the bulk four-velocity (which I consistently write as (GammaV, GammaV*V_drift) but within the code it often looks something like V_drift = (GammaV, GammaV*V_drift)), and other relativistic variables, feedback is appreciated.