One more refactor: originally I thought it would be a good idea to correct the whole four-velocity (GammaV, GammaV*V_drift). However, this has poor convergence properties, possibly due to it being non-trivial to properly set all the relevant quantities for the LTE distribution (since adjustments of GammaV affect n, the rest-frame density).
Instead, we go back to the formulation before, correcting (n, V_drift, T/m) where now V_drift in the relativistic case is the spatial component of the four-velocity u_i = GammaV*V_drift.

Then, in places where we need GammaV, we simply compute GammaV = sqrt(1 + |u_i|^2) (either at quadrature points or as a modal expansion).

This approach has quite good convergence properties in the correction routine and superior robustness, as it avoids issues of having to check whether V_drift < c. u_i is guaranteed to be well-defined, even in regions of strong gradients. **As an example of the superior robustness, the SR BGK test fails in main for Nx = 256, but now works beautifully with this approach**

These changes have a number of added benefits, including the fact that I can now define the pressure computation for p>1 in 2D and 3D (I no longer require the problematic basis_inv method).

Need to do more testing on GPUs and refactor the unit tests based on these changes, but this should be basically ready for prime time.