Proposed small refactor to restore the layering in computing collisional moments and make extensions to cross-species collisions more transparent.
The essential idea is that collision operators should *always* make the call they need to compute their needed moments. There should not be special cases or obfuscation of this fact. There is a BGK moment update just like there is an LBO moment update (and eventually an FPO moment update).

But to insure consistency when passing an input distribution function, there is a call like before that internally computes moments from an input distribution and then calls the new vm_species_lte_from_moms function, which wraps the relevant projection and correction calls.