REALLY NASTY BUG! RACE CONDITION.
During the GPU hackathon, we explored parallelizing bin_op multiplications, especially the conf-phase weak multiplies which have fairly non-trivial cost, over components in addition to cells.

Unfortunately, not enough testing was done before this was merged into the hackathon branch. The fundamental problem is that in many places in the code, to avoid allocating additional phase space arrays, we structure the calls to the bin_op method like this:
  gkyl_dg_mul_conf_phase_op_range(&up->conf_basis, &up->phase_basis,
    f_max, up->num_ratio, f_max, conf_range, phase_range);

Note that both the output and input phase-space arrays are the same, i.e., this is an in-place operation.

Parallelizing over components, **there is no guarantee of the order of operations and components which have already been updated in memory will affect other components**.

Currently we avoid race conditions on in-place operations by **both** only parallelizing over cells and making a temporary array, so that the thread sequentially constructs the whole temporary array without impacting the components of the array in memory for each cell.

I am reverting the parallelization over components. We cannot utilize this improvement by changing a core call. We will revisit the pieces of the algorithm which do similar operations to determine where we might derive benefit in the case when the output array is different from the input array, but in every case where we do an in-place operation, we cannot benefit from this parallelization strategy.