During the GPU hackathon we made a decision to save on computational cost by not computing the exponential pre-factor in the LTE distribution function under the logic that we always correct the density so the pre-factor is not needed.
However, further testing has revealed that the exponential pre-factor *is* useful because it increases the accuracy of the higher moments of the drift velocity and temperature and since the correction algorithm only formally converges in the cell average, we are restoring the computation of the exponential pre-factor.

Updating unit tests and getting rid of the use_last_converged flag since now that we have tracked down all the issues in the correction algorithm the only time this is utilized is during one time step in the BGK SR test where the correction routine genuinely diverges (and thus we only want to correct the density).