A major refactor of the comm system: cleaned up the public-facing API by removing all function pointer definitions from the public header, moving them to a private header. This cleans up the whole system considerably. Also split out the IO methods (read and write) out of the comm header into its own header. This is not any functionality but simply cleans up IO from Comm, though of course they are related in parallel.
For now, disabled send/recv as this needs proper design specially as
an internal state object needs to be opaquely returned. But depending
on the route we go we may not this. Though it seems it is best to have
this functionality for other reasons anyway.

NOTE: This reorg has not been done for NCCL comm! This will need to be
done too.