Refactor of the "PKPM penalization" in the momentum solve. While the 10 moment Roe solve appears to be much more robust for a number of problems, especially nonlinear plasma problems in 2D and 3D, it has limitations.
Firstly, the Roe solve will completely crap out at absorbing wall BCs because the mass density is zero on one side of the interface and one of the Roe variables involves 1/rho. Secondly the Roe solve seems to be poorly behaved for reflecting wall BCs due to the stagnation point where u_avg = 1/2 (u^+ + u^-) ~ 0. Finally, the Roe solve also fails in cases where the diagonals of the pressure tensor Pxx, Pyy, and Pzz are negative at the interface (even if they may be mostly positive within the cell).

Thus, I have refactored the penalization computation to pre-compute the total penalization of the momentum equation before you enter the hyper_dg loop. This pre-computation finds *both* the maximum speed for the Lax fluxes lambda_i = |u_i| + sqrt(3*Tii/m) and the total penalization term, which is either just the jump in the fluctuations found from the Roe solve, or is a simple Lax flux if any of the above conditions are met, which are simply assessed with a tolerance (if rho^+, rho^-, u_avg, Tii^+, or Tii^- are less than a tolerance, which works in the case they are identically 0 or problematically negative).

A bunch of regression tests that were failing with the new 10 moment Roe solve approach now work again including the pkpm_sheath and pkpm_wall tests. Will check that I did not break anything in multiple dimensions and also check GPU functionality.