Hopefully converging on some final polishing of the new penalization formulation. It turns out way back in commit 21216d2, while it is true that using u in the Roe solve was slightly more robust for 1D problems, I did not do enough aggressive testing of 2D and 3D (which I am doing now in preparation for the dg_10m branch to finally be merged).
The most robust multi-dimensional formulation evaluates rhou at quadrature points for the Roe solve (which makes sense; this is the conserved variable we are penalizing).

Also introduce some more sensible checking for negative density and temperature and thus switching to Lax fluxes. I am not yet certain how best to switch to Lax fluxes in the case of stagnation points. Right now what I consistently find is that the stagnation points are only an issue if u is supersonic, meaning the combination of u_avg and u^+/- > sqrt(3*Tii/m) roughly. I might try this as a more straightforward condition.