Currently, the way the implicit PKPM-EM coupling works is a modal DG generalization of the implicit fluid-EM coupling we utilize in the fluid solver (Wang et al. 2020 JCP). This modal generalization is well defined by multiplying the set of ODEs by test functions and integrating to make a linear system which can be inverted (analogous to our concept of weak division and the primitive variable corrections for the LBO).
However, for reasons that are not immediately clear to me, this inversion in the modal sense can fail. My immediate guess for *why* is that in cells in which the slopes are very steep (or higher order coefficients are large) and one can say the solution is oscillatory *within* the cell, the condition number for the matrix we are inverting is large.

As a test of this theory, or perhaps more accurately a workaround, I have implemented a nodal version of the implicit PKPM-EM coupling to test on problems where the implicit solve fails. This nodal version is really only accurate for p=1, because the conversion to the Gauss-Legendre basis via the Vandermonde matrix only integrates polynomials of order (2*(p+1) - 1), so 3 for p=1, 5 for p=2, and 7 for p=3, and we need to be able to integrate polynomials of order 3*p in the implicit solve (e.g., the u x B force is a linear solve because B has no sources, but the variation of B in the cell makes this part of the update require 3*p after multiplying by test functions). But this node-by-node implicit solve should always return a valid solve. Time to find out.