Adding the skeleton of a new equation object to handle photon transport. Right now this equation system is for a particular geometric configuration (1x, with 2v corresponding to the photon transport along the field line and the photon deflection off the field line, related to the radius of curvature of the field line). Still need some more infrastructure for computing |kpar| but will hopefully add that after iterating with Sasha C. on the best way to handle the various source terms (and thus the full scope of what we should pre-computed on the grid).
As a note: I am right now assuming the photon equation utilizes non-uniform grids in the way this branch has encoded them (with the cubic C^1 representation of the mapping). In this way, I can prevent obscene kernel proliferation and the cost is not extreme because the photon equation is relatively lightweight as is.