The mom-beach test in PKPM is fine, but with some of the changes to switch between Roe and Lax fluxes in the PKPM surface terms, the solver became slightly more brittle, causing the simulation to generate negative density near the right wall where the density is very low (20 orders of magnitude smaller than the left wall). Just increasing the density floor slightly from 1.0 to 10.0 (so the variation is only 19 orders of magnitude) fixes the numerical instability.
I believe this brittleness is caused by the fact that I am switching to Lax fluxes if the mass density is low in an absolute magnitude sense (and since this simulation is utilizing the actual electron mass, the mass density is 1.0e-29 at the right wall). It is possible the correct thing to do is to check the density, not the mass density, so that running with real units does not cause the system to aggressively switch to Lax fluxes.