Now that PKPM is being used for a variety of production calculations a few people have reported an issue when running with the implicit source split I generalized from our finite volume scheme where the linear solve will fail and it will throw the assert for status=false (meaning the matrix failed to be inverted in one or more of the cells in the domain).
As I documented in this commit in a prototyping branch: https://github.com/ammarhakim/gkylzero/commit/a60456d89917952efc5597ceecd7a86a0a4fcde6

One of the reasons this might be happening is that in cells where the slopes are very steep, the condition number of the matrix is very large. The fix in that branch was to do things nodally. I will bring this functionality over to production after more testing, but as an initial test of my theory, I am experimenting with switching from implicitly solving for du/dt to solving for dJ/dt (as we do in the finite volume scheme). Originally I thought it would be better to solve for du/dt and do a genuine solve on the primitive variables, but in recent weeks with the experimentation with L2 stability, I would now like to avoid weak divisions if possible (and we can avoid weak divisions for the non-relativistic source solve since the conserved variable is rhou, which is just the current density in the Ampere-Maxwell Law up to charge/mass factors).